+++ include/lexer.h
@@ -123,6 +123,9 @@ int				lx_word_start(void);
 t_token			*token_new(int type);
 void			token_del(t_token **tok);
 void			token_list_del(t_token **token);
+t_token			*token_dup(t_token *token);
+t_token			*token_list_dup(t_token *list);
+
 int				is_reserved_word(char *word);
 bool			is_operator_start(char c);
 bool			is_operator_part(char c);
@@ -130,6 +133,7 @@ bool			is_operator_next(char *ope, char c);
 bool			is_redir(t_token *token);
 int				get_operator_type(char *ope);
 char			*get_quote_string(enum e_quote_st quote);
+
 int				lx_line_insert_char(char c, int index);
 int				lx_line_del_char(int index);
 char			*get_prompt(void);
